+++ app/startup_other.go
@@ -1,0 +1,5 @@
+//go:build !windows
+
+package main
+
+func setStartOnLogin(string, bool) error { return nil }